Created internal.ts for shared private functions, updated toolrunner to call which(), prevent mock-task from double registering uncaughtException
