+++ node/index.ts
@@ -1,4 +1,3 @@
 /// <reference path="typings/index.d.ts" />
 
 import trm = require('./task');
-require('./extensions');
